// ui/src/app/core/ipc/gen/ModelRecommendation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ModelRecommendation = { name: string, label: string, description: string, downloadSize: bigint, estimatedMemory: bigint, installed: boolean, recommended: boolean, fit: string, };